requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
ijson==3.3.0
mdxpy==1.3.2
networkx==3.4.2
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
TM1py==2.0.4
urllib3==2.3.0
